requirements.txt
pandas~=1.5.3
numpy~=1.24.1
pyyaml
keras=2.11.0
matplotlib
tensorflow=2.11.0